Initial project structure / build setup
This commit provides the initial structure and build rules for the
java-time-client.

To start, this includes placeholder java and javatests code under the
com.google.time.client.base package. This package will contain the
common based classes needed for later commits.

Files have had licenses added using:
`addlicense -c "Google LLC" -l apache .`
... with the https://github.com/google/addlicense tool.

See README.md for details of the unusual directory structure.

Key to placeholder class naming:
+ {No prefix} : A class that is common to Java SE and Android variants
  of the library.
+ Android* : A class that only exists in the Android variant of the
  library.
+ JavaSe* : A class that only exists in the Java SE variant of the
  library.
+ Variant* : A class that exists in both the Java SE and Android
  variants, but with a variant-specific implementation.

The real classes that will replace the placeholders will not be named
like this. The names are purely to make it easier to understand. The
placeholders will be deleted as real code is added.