Initial base / test support code / helper scripts
This is an initial upload of code in com.google.time.client.base with
variants for Android / Java SE where needed, with tests and testing
support code.

This will form the base layer that the initial SNTP client will build
on. Functionality provided here is what a prototype of the SNTP client
suggests will be needed, but classes may be changed further as things
mature.

More packages may be added / split out later.