Switch to "client data minimization" by default
The original NTP specs suggest sending the client's instant in the NTP
request (optionally, with some randomization of the lower bits rather
than zero if the instant source has only millis precision).

https://datatracker.ietf.org/doc/html/draft-ietf-ntp-data-minimization-04
suggests this is an unnecessary "leak" of client information since the
information is not needed by the client.

This commit switches to sending a entirely random value by default. The
new behavior can be turned off if required.