Add errorprone suppressions for newer API refs
Google tooling picks up on the use of newer java.time APIs only present
from API 26. I've not found a way to make the tooling aware that this is
ok: The methods are clearly marked as @RequiresApi and so should not be
used (and, I assume, won't cause an issue at runtime unless the method
is resolved in preparation for a call), so I think suppressing these
warnings should be safe and should consistitute a no-op outside of
Google.