Reduce the severity of unusual / unexpected timeout behavior
Before this commit, ClusteredServiceOperation treats the following
scenario as a fatal error (IllegalStateException):

1) Clustered operation has N seconds to execute and passes that to the
   service operation.
2) Service operation says reports it ran out of time in < N seconds.

In the SNTP implementation, it looks like this can happen when the
SntpClientOperation is given less time to complete than it would usually
wait for a UDP response, e.g. if it would usually wait 5 seconds for a
response but it is only given 3 seconds. In these cases, rather than
setting a socket timeout of 5 seconds, it sets 3 seconds and relies on
the DatagramSocket, and ultimately its implementation and maybe even the
kernel to enforce this timeout. When a SocketTimeoutException is thrown
it reports this as "allowed time exceeded".

What appears to be possible is that the SocketTimeoutException can be
thrown "early", triggering the ClusteredServiceOperation
IllegalStateException. Speculating, this might be due to a different
clock being used or (most likely, based on prior, hazily remembered,
experience), the HZ kernel setting in Linux, which I believe can cause
timeouts to trigger early.

In any case, treating this case as fatal is causing problems in some
environments / scenarios, so this commit reduces the severity to just
logging and pretending the socket timeout worked as intended.