Support elapsed realtime clock on Android
Add better support for converting to/from an elapsed realtime clock
value on Android. A recent google3 internal user found themselves having
to use Ticks.fromTickerValue() in production code, and generally use of
that method outside of tests and Ticker impls is discouraged. The
fromTickerValue() has been made package-private and Ticker.createTicks()
should be used instead. This is a breaking change.

The elapsed realtime clock is an Androidism, so the PlatformTicker for
Android has been changed to extend a new AndroidTicker base class that
allows (optional) conversion to/from elapsed realtime long values and
Ticks objects.

PlatformTickerTest has been made platform-specific and tests added for
the new AndroidTicker methods.

LinearAdjustmentTicker has been deleted as it's not used and recent
tests on Android suggest that a trivial linear adjustment is not
sufficient to account for clock inaccuracies on Android devices. This is
a breaking change, but not one that I expect to break anybody.