Adjust scripts for changes to bazel
bazel appears to now use a sandbox by default. This is deliberate and
reasonable, but plays badly with this project's use of symlinks. The
introduction of java-time-client.bazelrc is to disable the sandboxing.

I don't currently have a good alternative answer for sharing code across
two very related projects (javase / android) without duplication where
any class could be specialized. It was a deliberate decision not to add
interfaces/factories, etc. or other normal approaches to cross-platform.
See the project README.md for info.